[upd] IAM SubModule - Allow alternative trusted role arn for testing
To automate testing of the Terraform modules, we need to be able to pass
in a non-production role arn to the IAM submodule. This change allows
the automated testing to pass in a new trusted role arn. This parameter
should not be used by any users and defaults to null in the root module.

On branch release-0.12.0
Changes to be committed:
	modified:   README.md
	modified:   main.tf
	modified:   modules/aws-anyscale-iam/README.md
	modified:   modules/aws-anyscale-iam/iam-policies-data.tf
	modified:   modules/aws-anyscale-iam/variables.tf
	modified:   modules/aws-anyscale-vpc/main.tf
	modified:   variables.tf