[upd] CHANGELOG for release-0.13.0
[upd] IAM SubModule - Allow alternative trusted role arn for testing
To automate testing of the Terraform modules, we need to be able to pass
in a non-production role arn to the IAM submodule. This change allows
the automated testing to pass in a new trusted role arn. This parameter
should not be used by any users and defaults to null in the root module.

[upd] AWS Terraform Provider to ~> 5.0
Includes updates to VPC module to support new provider and deprecated
settings for EIP.

On branch release-0.13.0
Changes to be committed:
	modified:   CHANGELOG.md