[upd] aws-anyscale-iam submodule - Add CloudWatch IAM policy
This update includes the addition of a CloudWatch IAM policy to the
aws-anyscale-iam module.  This policy is required for the the new
monitoring functionality.

This policy is not enabled by default.  To enable it, set the variable
`create_cluster_node_cloudwatch_policy = true` in the root module.

There was also some general cleanup of variable descriptions and tested
with the latest AWS terraform provider.

On branch brent/iam-cloudwatch
Changes to be committed:
	modified:   README.md
	modified:   examples/anyscale-v2-commonname/README.md
	modified:   examples/anyscale-v2-commonname/main.tf
	modified:   examples/anyscale-v2-commonname/variables.tf
	modified:   main.tf
	modified:   modules/aws-anyscale-efs/README.md
	modified:   modules/aws-anyscale-iam/README.md
	modified:   modules/aws-anyscale-iam/iam-policies-data.tf
	modified:   modules/aws-anyscale-iam/main.tf
	modified:   modules/aws-anyscale-iam/variables.tf
	modified:   modules/aws-anyscale-s3-policy/README.md
	modified:   modules/aws-anyscale-s3-policy/examples/README.md
	modified:   modules/aws-anyscale-s3/README.md
	modified:   modules/aws-anyscale-securitygroups/README.md
	modified:   modules/aws-anyscale-vpc/README.md
	modified:   variables.tf