[upd] root module variables - general cleanup of descriptions
This is a pass at cleaning up the descriptions of the root module's
variables.tf file to make it easier to read and understand. Especially
with the way tfdocs renders the descriptions, it's important to have
them be as clear as possible.

General validation with the latest AWS provider.

On branch brent/variable-cleanup
Changes to be committed:
	modified:   README.md
	modified:   modules/aws-anyscale-efs/README.md
	modified:   modules/aws-anyscale-iam/README.md
	modified:   modules/aws-anyscale-s3-policy/README.md
	modified:   modules/aws-anyscale-s3-policy/examples/README.md
	modified:   modules/aws-anyscale-s3/README.md
	modified:   modules/aws-anyscale-securitygroups/README.md
	modified:   modules/aws-anyscale-vpc/README.md
	modified:   variables.tf