[upd] Root Module - add new variable for IAM Policy ARNs
Added new variable: anyscale_cluster_node_managed_policy_arns
This is passed to the IAM Submodule variable with the same name.

This allows the addition of IAM Policies to the cluster node role via IAM policy ARNs.
Added example showing how this can be done to anyscale-v2-commonname example.

Changes to be committed:
	modified:   CHANGELOG.md
	modified:   README.md
	modified:   examples/anyscale-v2-commonname/README.md
	modified:   examples/anyscale-v2-commonname/main.tf
	modified:   main.tf
	modified:   variables.tf