[upd] aws-anyscale-iam: Add Optional SecretsManager IAM policy
This commit adds a new IAM policy for SecretsManager.  This policy allows
passing in a new parameter with a list of secrets - as well as a KMS key to use (optional).
This policy is then assigned to the Cluster Node IAM role. This allows the
cluster to read the secret which is then used for BYOD (Bring Your Own Docker) images
that are in private, password protected repositories.

Additional updates include a new example named kitchensink (WIP), updates to
the latest Terraform and AWS Provider versions. There are also updates to
the unit tests with the plan to automate building the resources,
deploying an Anyscale Cloud, using Functional Verification, and then
destroying the resources as part of an automated github action (also a WIP).

On branch brent/iam-secrets
Changes to be committed:
	modified:   .tflint.hcl
	modified:   CHANGELOG.md
	modified:   README.md
	deleted:    examples/anyscale-v2-createendpoints/README.md
	deleted:    examples/anyscale-v2-createendpoints/main.tf
	deleted:    examples/anyscale-v2-createendpoints/outputs.tf
	modified:   examples/anyscale-v2-existing-vpc/main.tf
	new file:   examples/anyscale-v2-kitchensink/README.md
	new file:   examples/anyscale-v2-kitchensink/main.tf
	new file:   examples/anyscale-v2-kitchensink/outputs.tf
	renamed:    examples/anyscale-v2-createendpoints/variables.tf -> examples/anyscale-v2-kitchensink/variables.tf
	renamed:    examples/anyscale-v2-createendpoints/versions.tf -> examples/anyscale-v2-kitchensink/versions.tf
	modified:   main.tf
	modified:   modules/aws-anyscale-efs/README.md
	modified:   modules/aws-anyscale-iam/README.md
	new file:   modules/aws-anyscale-iam/cluster_node-secretsmanager-getsecret-kms.tmpl
	new file:   modules/aws-anyscale-iam/cluster_node-secretsmanager-getsecret.tmpl
	modified:   modules/aws-anyscale-iam/iam-policies-data.tf
	modified:   modules/aws-anyscale-iam/main.tf
	modified:   modules/aws-anyscale-iam/test/anyscale-test/main.tf
	modified:   modules/aws-anyscale-iam/test/anyscale-test/outputs.tf
	modified:   modules/aws-anyscale-iam/variables.tf
	modified:   modules/aws-anyscale-s3-policy/README.md
	modified:   modules/aws-anyscale-s3-policy/examples/README.md
	modified:   modules/aws-anyscale-s3/README.md
	modified:   modules/aws-anyscale-securitygroups/README.md
	modified:   modules/aws-anyscale-vpc/README.md
	new file:   test/REQUIREMENTS.txt
	new file:   test/test_cloud_register_manual.py
	modified:   variables.tf