[upd] kitchensink example and e2e test AWS Provider versions
The AWS Provider version 5.17.0 has a bug which causes s3 bucket deletion
via terraform to fail. This commit updates the kitchensink example and the
e2e unit test to use the AWS Provider version 5.16.x.

It also updates pre-commit-config to not update to the lates 5.0 version.

Changes to be committed:
	modified:   .pre-commit-config.yaml
	modified:   CHANGELOG.md
	modified:   examples/anyscale-v2-kitchensink/README.md
	modified:   examples/anyscale-v2-kitchensink/versions.tf
	modified:   test/anyscale-v2-e2e-public-test/versions.tf
	modified:   test/test_cloud_register_manual.py