fix: Existing S3 bucket functionality
Using an existing S3 bucket via the example `anyscale-v2-existing-s3` was
broken with the latest versions of Terraform due to an invalid IAM policy
reference to the S3 bucket name. This fixes that - and updates the example
to provide a working output command that can be used to register an Anyscale
Cloud.

On branch brent/fix-existings3
Changes to be committed:
	modified:   CHANGELOG.md
	modified:   README.md
	modified:   examples/anyscale-v2-existing-s3/README.md
	modified:   examples/anyscale-v2-existing-s3/main.tf
	modified:   examples/anyscale-v2-existing-s3/outputs.tf
	modified:   main.tf