upd: EKS Cluster and Nodegroups modules
The EKS Cluster does not need the Anyscale Security Group so that
was removed.

The EKS Node Groups do require the Anyscale Security Group, but
made those default to `null` in the variables so the nodegroups
can be created without the Anyscale Security Group (if desired).

Changes to be committed:
	modified:   modules/aws-anyscale-eks-cluster/README.md
	modified:   modules/aws-anyscale-eks-cluster/variables.tf
	modified:   modules/aws-anyscale-eks-nodegroups/README.md
	modified:   modules/aws-anyscale-eks-nodegroups/main.tf
	modified:   modules/aws-anyscale-eks-nodegroups/test/anyscale-test/main.tf
	modified:   modules/aws-anyscale-eks-nodegroups/variables.tf