upd: Make port 22 optional in security group config (#64)
With recent updates to the Anyscale CLI, VSCode, and Cursor plugins, port 22 is no longer required.

This requires the Anyscale CLI to be updated to v0.26.30 or later and the VSCode and Cursor plugins to be updated to v0.2.33 or later.

- Remove default port 22 from security group config
- Add new variable `security_group_enable_ssh_access` with a default of false to enable SSH access if required.
- Update the kitchensink example to use the new variable and enable port 22.

On branch brent/remove-port22-default
Changes to be committed:
	modified:   README.md
	modified:   examples/anyscale-v2-kitchensink/README.md
	modified:   examples/anyscale-v2-kitchensink/main.tf
	modified:   examples/anyscale-v2-kitchensink/variables.tf
	modified:   main.tf
	modified:   variables.tf