upd: IAM Role descriptions
Remove the region from the IAM Role descriptions.
With the new AWS provider, the call to the data source to return the region introduced
a deprecation warning. As this is rarely required, it is being removed and the region
will not be part of the IAM Role descriptions.

Changes to be committed:
	modified:   modules/aws-anyscale-iam/README.md
	modified:   modules/aws-anyscale-iam/eks-iam-main.tf
	modified:   modules/aws-anyscale-iam/main.tf