add: Update Security Group configuration to allow either CIDR or Existing Security Group configurations.
This change allows for the use of either CIDR or Existing Security Group configurations
for the security group ingress rules. This is useful for customers who have existing
security groups they want to use to grant access to the Anyscale data plane.

This also cleans up commented-out code from older releases.

On branch brent/update-security-group-config
Changes to be committed:
	modified:   CHANGELOG.md
	modified:   README.md
	modified:   main.tf
	modified:   modules/aws-anyscale-securitygroups/examples/main.tf
	modified:   modules/aws-anyscale-securitygroups/examples/outputs.tf
	modified:   modules/aws-anyscale-securitygroups/main.tf
	modified:   modules/aws-anyscale-securitygroups/variables.tf
	modified:   variables.tf