upd: Cloud Resource example outputs (#82)
* upd: commonname example outputs and README

Added the common prefix to the output for the name in the yaml file
Ensured that pre-commit executed for README updates.

On branch brent/update-example-readmes
Changes to be committed:
	modified:   README.md
	modified:   outputs.tf

* upd: CommonName and privatesubnets outputs and README

Added the external id to the outputs for the yaml file
Ensured that pre-commit executed for README updates.

On branch brent/update-example-readmes
Changes to be committed:
	modified:   anyscale-v2-commonname/outputs.tf
	modified:   anyscale-v2-privatesubnets/README.md
	modified:   anyscale-v2-privatesubnets/outputs.tf